feat: show personal info in identity LOC user view.
logion-network/logion-internal#578
